Working on Allowing user to save an Edit
I have created a function in the documents controller. Working on
allowing the user to save edits. I am creating a form with various
inputs of the old document like the id the title and everything else
stays  the same the only new info is the edited document. I then post
that to the new function and then save the doc that way.