User's can now edit and save changes to document.
I have it so that the user can now edit the document. The user will
select the document in the grid they will be then presented with a
document preview. Then the user will then select the edit document
button. From there the user can edit the document then save.

The issue I am having now is I’m get some error when selecting
different documents in the grid has something to do with signature.