Password are now md5() encrypted
Added md5() encryption to the passwords when a user is newly created
the password is md5() encrypted and then stored in the database. When
the users logs back on the password is md5() hashed and checked against
the md5() hash in the database if match then alls is good and user is
permitted to log on.