Fixed issue where user was seeing wrong stuff
The user was always seeing the documents that need to be approved in
every tab that they chose. I fixed this by adding this under the if
statement in the main.php view that checks if the user is an admin. The
project is pretty much done pending further testing.